+++ packages/mint-components/src/components/sqm-checkbox-field/CheckboxField.stories.tsx
@@ -25,7 +25,7 @@ const defaultProps: CheckboxFieldViewProps = {
     checked: false,
   },
   content: {
-    checkboxName: "agree",
+    checkboxName: "terms",
     checkboxLabel: "By signing up you agree to the {labelLink}",
     checkboxLabelLink: "https://example.com",
     checkboxLabelLinkText: "Terms and Conditions",