+++ packages/mint-components/src/components/tax-and-cash/sqm-tax-and-cash-dashboard/sqm-tax-and-cash-dashboard-view.tsx
@@ -754,7 +754,7 @@ export const TaxAndCashDashboardView = (props: TaxAndCashDashboardProps) => {
                 {text.indirectTaxInfoSectionHeader}
                 <sl-tooltip
                   trigger="hover"
-                  placement="top"
+                  placement="right"
                   content={text.indirectTaxTooltipSupport}
                   class={sheet.classes.TooltipContainer}
                 >